Stop using deprecated removeField method
taglib deprecated removeField internally in [1], but did not add the
deprecation warning until about four years later. The original bug
report [2] and pull request commit both mention not wanting to change
the original method's behaviour, but this was done anyway (perhaps by
mistake) in [3].

With that change, removeField(type) will not remove all tags of the
given type anymore, as the default for value has changed from
String::null to String().

This commit replaces explicit calls to removeField(type) with
removeFields(type), which has the correct behaviour. In writeXiph,
removeField is removed entirely, as addField(key, value) will replace
the tag by default.

[1] https://github.com/taglib/taglib/pull/681
[2] https://github.com/taglib/taglib/issues/651
[3] https://github.com/taglib/taglib/commit/c05fa78406fd8ce7382a11c1f63a17c4bfbe83fa